+++ test-project-js/export-folder/reexported.js
@@ -1,1 +1,4 @@
 export const reexportedConst = 1;
+
+const reexportedConst2 = 2;
+export default reexportedConst2;