+++ test-project-js/import-folder/import-after-file.js
@@ -1,0 +1,4 @@
+import './one.css';
+import { one } from '../export-folder/one';
+
+console.log(one);