+++ test-project-js/re-exports/reexports2.js
@@ -1,1 +1,4 @@
 export { reexportedConst } from '../export-folder/reexported';
+
+import reexportedConst2 from '../export-folder/reexported';
+export { reexportedConst2 };